Changed openMP for loops from size_t to int to compile on Windows.

+++ include/dkm_parallel.hpp
@@ -35,7 +35,7 @@ std::vector<T> closest_distance_parallel(
 	const std::vector<std::array<T, N>>& means, const std::vector<std::array<T, N>>& data) {
 	std::vector<T> distances(data.size(), T());
 	#pragma omp parallel for
-	for (size_t i = 0; i < data.size(); ++i) {
+	for (int i = 0; i < data.size(); ++i) {
 		T closest = distance_squared(data[i], means[0]);
 		for (const auto& m : means) {
 			T distance = distance_squared(data[i], m);
@@ -91,7 +91,7 @@ std::vector<uint32_t> calculate_clusters_parallel(
 	const std::vector<std::array<T, N>>& data, const std::vector<std::array<T, N>>& means) {
 	std::vector<uint32_t> clusters(data.size(), 0);
 	#pragma omp parallel for
-	for (size_t i = 0; i < data.size(); ++i) {
+	for (int i = 0; i < data.size(); ++i) {
 		clusters[i] = closest_mean(data[i], means);
 	}
 	return clusters;
